+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.5.13 (SNAPSHOT)
 * Add tar type to package target
 * Fix mocha type test* to run all tests in subdirectories
+* Add buster test option
 
 ### 0.5.12
 * Fix package zip type opts to make it work on RHEL other than Ubuntu and OS X